requirements.txt
fastapi
uvicorn
pandas
python-dotenv
httpx
google-generativeai
python-multipart
